[REFACTOR]: Separate useEventMouse hook
